HW-2 불필요한 소스 삭제 및 구조 변경
