+++ class12/manjee-study/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/model/MovieMeta.kt
@@ -1,7 +1,5 @@
 package com.example.architecturestudy.data.model
 
-import com.example.architecturestudy.data.model.MovieData
-
 data class MovieMeta(
     val display: Int,
     val items: List<MovieData>,
