+++ class12/manjee-study/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/repository/MovieRepository.kt
@@ -1,7 +1,11 @@
 package com.example.architecturestudy.data.repository
 
-import com.example.architecturestudy.data.source.remote.RemoteCallback
+import com.example.architecturestudy.data.model.MovieData
 
 interface MovieRepository {
-    fun remoteSearchMovie(title: String, searchMovieCallback: RemoteCallback)
+    fun remoteSearchMovie(
+        title: String,
+        success: (List<MovieData>) -> Unit,
+        fail: (Throwable) -> Unit
+    )
 }
