+++ class12/manjee-study/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/repository/MovieRespositoryImpl.kt
@@ -1,10 +1,33 @@
 package com.example.architecturestudy.data.repository
 
+import com.example.architecturestudy.data.model.MovieData
+import com.example.architecturestudy.data.model.MovieMeta
 import com.example.architecturestudy.data.source.remote.MovieRemoteService
-import com.example.architecturestudy.data.source.remote.RemoteCallback
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class MovieRespositoryImpl : MovieRepository {
-    override fun remoteSearchMovie(title: String, searchMovieCallback: RemoteCallback) {
-        MovieRemoteService.movieApiService.searchMovie(title).enqueue(searchMovieCallback)
+    override fun remoteSearchMovie(
+        title: String,
+        success: (List<MovieData>) -> Unit,
+        fail: (Throwable) -> Unit
+    ) {
+        MovieRemoteService.movieApiService.searchMovie(title).enqueue(object : Callback<MovieMeta> {
+            override fun onFailure(call: Call<MovieMeta>, t: Throwable) {
+                fail(t)
+            }
+
+            override fun onResponse(
+                call: Call<MovieMeta>,
+                response: Response<MovieMeta>
+            ) {
+                val metaData = response.body()
+                metaData?.let {
+                    success(it.items)
+                }
+            }
+
+        })
     }
 }
