--- class12/manjee-study/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/source/remote/RemoteCallback.kt
@@ -1,6 +1,0 @@
-package com.example.architecturestudy.data.source.remote
-
-import com.example.architecturestudy.data.model.MovieMeta
-import retrofit2.Callback
-
-interface RemoteCallback : Callback<MovieMeta>
